refactor: remove channel-guardian-store.ts re-export shim
