+++ .github/CODEOWNERS
@@ -104,7 +104,6 @@ assistant/src/tools/tool-approval-handler.ts @noanflaherty
 assistant/src/security/tool-approval-digest.ts @noanflaherty
 assistant/src/permissions/ @noanflaherty
 assistant/src/memory/guardian-* @noanflaherty
-assistant/src/memory/channel-guardian-store.ts @noanflaherty
 assistant/src/memory/scoped-approval-grants.ts @noanflaherty
 assistant/src/memory/canonical-guardian-store.ts @noanflaherty
 assistant/src/calls/guardian-*.ts @noanflaherty
