+++ assistant/src/daemon/handlers/config-inbox.ts
@@ -3,15 +3,15 @@ import * as net from "node:net";
 import type { ChannelId } from "../../channels/types.js";
 import { isChannelId, isInterfaceId } from "../../channels/types.js";
 import { getGatewayInternalBaseUrl } from "../../config/env.js";
-import type { GuardianApprovalRequest } from "../../memory/channel-guardian-store.js";
-import {
-  type ApprovalRequestStatus,
-  listPendingApprovalRequests,
-  resolveApprovalRequest,
-} from "../../memory/channel-guardian-store.js";
 import { addMessage, getMessages } from "../../memory/conversation-crud.js";
 import { getLatestStoredPayload } from "../../memory/delivery-crud.js";
 import { getBindingByConversation } from "../../memory/external-conversation-store.js";
+import {
+  type ApprovalRequestStatus,
+  type GuardianApprovalRequest,
+  listPendingApprovalRequests,
+  resolveApprovalRequest,
+} from "../../memory/guardian-approvals.js";
 import { DAEMON_INTERNAL_ASSISTANT_ID } from "../../runtime/assistant-scope.js";
 import { mintDaemonDeliveryToken } from "../../runtime/auth/token-service.js";
 import { deliverChannelReply } from "../../runtime/gateway-client.js";
